boshqarma/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cloudinary==1.28.0
cryptography==35.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.1.11
Django==3.1.3
django-allauth==0.44.0
django-cleanup==5.2.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.10.1
django-magic==0.5a0
djangorestframework==3.12.4
idna==3.3
oauthlib==3.1.1
Pillow==8.4.0
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
